refactor(api): extract mapWithConcurrency into shared lib
