+++ apps/api/tests/lib/concurrency.test.ts
@@ -1,0 +1,25 @@
+import { describe, it, expect } from 'vitest';
+import { mapWithConcurrency } from '../../src/lib/concurrency.js';
+
+describe('mapWithConcurrency', () => {
+  it('maps all items preserving order', async () => {
+    const results = await mapWithConcurrency([1, 2, 3, 4], 2, async (n) => n * 10);
+    expect(results).toEqual([10, 20, 30, 40]);
+  });
+
+  it('never runs more than `limit` tasks at once', async () => {
+    let inFlight = 0;
+    let maxInFlight = 0;
+    await mapWithConcurrency([1, 2, 3, 4, 5, 6], 2, async () => {
+      inFlight++;
+      maxInFlight = Math.max(maxInFlight, inFlight);
+      await new Promise((r) => setTimeout(r, 10));
+      inFlight--;
+    });
+    expect(maxInFlight).toBe(2);
+  });
+
+  it('handles empty input', async () => {
+    expect(await mapWithConcurrency([], 5, async (x) => x)).toEqual([]);
+  });
+});
